multiple delete functionality added

+++ backend/controllers/task/deleteSelectedTask.js
@@ -4,10 +4,13 @@ async function deleteSelectedTask(req, res) {
   try {
     const { selectedId } = req.body;
 
+    if (!selectedId) {
+      return res.status(404).json({message:"No body found for deleting "})
+    }
     if (!Array.isArray(selectedId) || selectedId.length === 0) {
       return res
         .status(400)
-        .json({ message: "No task IDs provided for deletion." });
+        .json({ message: "No task IDs provided for deletion.", selectedId });
     }
 
     const selectionDeleted = await TaskModel.deleteMany({
@@ -20,6 +23,8 @@ async function deleteSelectedTask(req, res) {
 
     res.status(200).json({
       message: `${selectionDeleted.deletedCount} tasks deleted successfully.`,
+      selectionDeleted,
+      selectedId,
     });
   } catch (error) {
     console.log("Internal server Error during deleting tasks", error);
